Update NoseconeTipTemp.m
Everything seems to line up, but something is going wrong in terms of it not reaching tolerance convergence, and I cannot figure out where. However, majority of the code is set up to be used for the nosecone tip, and will be set up brand new for the fin LE until issue is found.