Combined the two Monte Carlo scripts and defined as functions
Took the two Monte Carlo scripts and combined it into one. Will likely move into using the classes in individual files for better expandability and application. Also parametrised number of sims and the goal. via an index choice. Would like to further parametrise with the types and sizes of perturbations.